fix surgery page blocks, fix affiliate hours

+++ resources/views/livewire/admin/surgery/block/create-edit.blade.php
@@ -138,46 +138,48 @@
                                                 </div>
                                             </div>
 
-                                            <div class="col-md-12">
-                                                <div class="row">
-                                                    <div class="col-md-12">
-                                                        <div class="row mb-3">
-                                                            <div class="col-md-12">
-                                                                <div class="tab-content">
-                                                                    <div
-                                                                        class="multilang-content tab-pane fade active show ">
-                                                                        <div class="form-group mb-1">
-                                                                            <label>{{ __('admin.image') }}</label>
-                                                                            <input type="file" wire:model="image"
-                                                                                class="form-control">
+                                            @if(in_array($this->block->block ?? '', ['Inpatient', 'conditions']) && $this->block->key ?? '' == 'title' ? false : true)
+                                                <div class="col-md-12">
+                                                    <div class="row">
+                                                        <div class="col-md-12">
+                                                            <div class="row mb-3">
+                                                                <div class="col-md-12">
+                                                                    <div class="tab-content">
+                                                                        <div
+                                                                            class="multilang-content tab-pane fade active show ">
+                                                                            <div class="form-group mb-1">
+                                                                                <label>{{ __('admin.image') }}</label>
+                                                                                <input type="file" wire:model="image"
+                                                                                    class="form-control">
+                                                                            </div>
                                                                         </div>
                                                                     </div>
-                                                                </div>
-                                                                @error('image')
-                                                                    <div class="mt-1 text-danger ajaxError">
-                                                                        {{ $message }}
-                                                                    </div>
-                                                                @enderror
+                                                                    @error('image')
+                                                                        <div class="mt-1 text-danger ajaxError">
+                                                                            {{ $message }}
+                                                                        </div>
+                                                                    @enderror
 
-                                                                @if ($this->imageTemporary)
-                                                                    <div class="flex">
-                                                                        <img src="{{ $this->imageTemporary }}"
+                                                                    @if ($this->imageTemporary)
+                                                                        <div class="flex">
+                                                                            <img src="{{ $this->imageTemporary }}"
+                                                                                width="60">
+                                                                            <a wire:click="deleteImage()"
+                                                                                style="cursor: pointer;">
+                                                                                <i class="ti-close font-weight-bold mr-2"></i>
+                                                                                {{ __('admin.delete_image') }}
+                                                                            </a>
+                                                                        </div>
+                                                                    @elseif(!empty($this->block->image))
+                                                                        <img src="{{ $this->block->imageUrl }}"
                                                                             width="60">
-                                                                        <a wire:click="deleteImage()"
-                                                                            style="cursor: pointer;">
-                                                                            <i class="ti-close font-weight-bold mr-2"></i>
-                                                                            {{ __('admin.delete_image') }}
-                                                                        </a>
-                                                                    </div>
-                                                                @elseif(!empty($this->block->image))
-                                                                    <img src="{{ $this->block->imageUrl }}"
-                                                                        width="60">
-                                                                @endif
+                                                                    @endif
+                                                                </div>
                                                             </div>
                                                         </div>
                                                     </div>
                                                 </div>
-                                            </div>
+                                            @endif
                                         @endif
 
                                         @if(!empty($this->block->block) && $this->block->block == '3d')
